optimize memory; add friend debug commands. tweak charger parameters to 4.2V

+++ src/mic.h
@@ -2,7 +2,7 @@
 
 #define NUM_RX_SAMPLES 512
 extern int32_t rx_samples[];
-extern int32_t rx_savebuf[];
+extern uint16_t rx_savebuf[];
 extern uint32_t mic_return[MIC_SAMPLE_DEPTH];
 
 extern uint32_t rx_int_count;
@@ -13,4 +13,4 @@ void analogUpdateTemperature(void);
 int32_t analogReadTemperature(void);
 
 void analogUpdateMic(void);
-uint8_t *analogReadMic(void);
+uint16_t *analogReadMic(void);
